Reorganized damage type system, for bloodbath/recklessness.

+++ battlenet.cpp
@@ -221,7 +221,7 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 691: input->setText(qsprint(1870)); break;
 				case 700: input->setText(qsprint(2033)); break;
 				case 706: input->setText(qsprint(2150)); break;
-				default:  input->setText(qsprint((float)(2033.0 * pow(1.0098, itemlvl - 700)))); break;
+				default:  input->setText(qsprint((int)(2033.0 * pow(1.0098, itemlvl - 700)))); break;
 				}
 			}
 			if (itemid == 113983){
@@ -233,7 +233,7 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 691: input->setText(qsprint(167)); break;
 				case 700: input->setText(qsprint(181)); break;
 				case 706: input->setText(qsprint(192)); break;
-				default:  input->setText(qsprint((float)(181.0 * pow(1.0098, itemlvl - 700)))); break;
+				default:  input->setText(qsprint((int)(181.0 * pow(1.0098, itemlvl - 700)))); break;
 				}
 			}
 			if (itemid == 119193){
@@ -245,14 +245,14 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 691: input->setText(qsprint(2439)); break;
 				case 700: input->setText(qsprint(2652)); break;
 				case 706: input->setText(qsprint(2804)); break;
-				default:  input->setText(qsprint((float)(2652.0 * pow(1.0098, itemlvl - 700)))); break;
+				default:  input->setText(qsprint((int)(2652.0 * pow(1.0098, itemlvl - 700)))); break;
 				}
 			}
 			if (itemid == 118882){
 				combo->setCurrentIndex(4);
 				switch (itemlvl){
 				case 665: input->setText(qsprint(2200)); break;
-				default:  input->setText(qsprint((float)(2200.0 * pow(1.0098, itemlvl - 665)))); break;
+				default:  input->setText(qsprint((int)(2200.0 * pow(1.0098, itemlvl - 665)))); break;
 				}
 			}
 			if (itemid == 115159 || itemid == 119936 || itemid == 111232 || itemid == 115759 || itemid == 125041 || itemid == 125518 || itemid == 126632 || itemid == 126155 || itemid == 124867 || itemid == 125344 || itemid == 125981 || itemid == 126458){
@@ -265,7 +265,7 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 680: input->setText(qsprint(798)); break;
 				case 700: input->setText(qsprint(1057)); break;
 				case 705: input->setText(qsprint(1108)); break;
-				default:  input->setText(qsprint((float)(728.0 * pow(1.0098, itemlvl-660)))); break;
+				default:  input->setText(qsprint((int)(728.0 * pow(1.0098, itemlvl-660)))); break;
 				}
 			}
 			if (itemid == 115160 || itemid == 119937 || itemid == 111233 || itemid == 115760 || itemid == 125042 || itemid == 125519 || itemid == 126633 || itemid == 126156 || itemid == 124868 || itemid == 125345 || itemid == 125982 || itemid == 126459){
@@ -278,7 +278,7 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 680: input->setText(qsprint(951)); break;
 				case 700: input->setText(qsprint(1259)); break;
 				case 705: input->setText(qsprint(1319)); break;
-				default:  input->setText(qsprint((float)(867.0 * pow(1.0098, itemlvl - 660)))); break;
+				default:  input->setText(qsprint((int)(867.0 * pow(1.0098, itemlvl - 660)))); break;
 				}
 			}
 			if (itemid == 113645){
@@ -290,14 +290,14 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 676: input->setText(qsprint(2122)); break;
 				case 685: input->setText(qsprint(2304)); break;
 				case 691: input->setText(qsprint(2439)); break;
-				default:  input->setText(qsprint((float)(2304.0 * pow(1.0098, itemlvl - 685)))); break;
+				default:  input->setText(qsprint((int)(2304.0 * pow(1.0098, itemlvl - 685)))); break;
 				}
 			}
 			if (itemid == 114613){
 				combo->setCurrentIndex(8);
 				switch (itemlvl){
 				case 655: input->setText(qsprint(1743)); break;
-				default:  input->setText(qsprint((float)(1743.0 * pow(1.0098, itemlvl - 655)))); break;
+				default:  input->setText(qsprint((int)(1743.0 * pow(1.0098, itemlvl - 655)))); break;
 				}
 			}
 			if (itemid == 109262 || (itemid >= 122601 && itemid <= 122604)){
@@ -308,7 +308,7 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 655: input->setText(qsprint(1069)); break;
 				case 670: input->setText(qsprint(1229)); break;
 				case 685: input->setText(qsprint(1414)); break;
-				default:  input->setText(qsprint((float)(1414.0 * pow(1.0098, itemlvl - 685)))); break;
+				default:  input->setText(qsprint((int)(1414.0 * pow(1.0098, itemlvl - 685)))); break;
 				}
 			}
 			if (itemid == 112318){
@@ -320,7 +320,7 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 685: input->setText(qsprint(2120)); break;
 				case 700: input->setText(qsprint(2440)); break;
 				case 715: input->setText(qsprint(2804)); break;
-				default:  input->setText(qsprint((float)(2804.0 * pow(1.0098, itemlvl - 715)))); break;
+				default:  input->setText(qsprint((int)(2804.0 * pow(1.0098, itemlvl - 715)))); break;
 				}
 			}
 			if (itemid == 116292){
@@ -328,7 +328,7 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				switch (itemlvl){
 				case 640: input->setText(qsprint(1517)); break;
 				case 646: input->setText(qsprint(1604)); break;
-				default:  input->setText(qsprint((float)(1517.0 * pow(1.0098, itemlvl - 640)))); break;
+				default:  input->setText(qsprint((int)(1517.0 * pow(1.0098, itemlvl - 640)))); break;
 				}
 			}
 			if (itemid == 124523){
@@ -340,7 +340,7 @@ void gic::import_player(std::string& realm, std::string& name, std::string& regi
 				case 726: input->setText(qsprint(267)); break;
 				case 735: input->setText(qsprint(290)); break;
 				case 741: input->setText(qsprint(307)); break;
-				default:  input->setText(qsprint((float)(290.0 * pow(1.0098, itemlvl - 735)))); break;
+				default:  input->setText(qsprint((int)(290.0 * pow(1.0098, itemlvl - 735)))); break;
 				}
 			}
 		}
